+++ PhotoshopAPI/src/LayeredFile/LayeredFile.h
@@ -71,7 +71,7 @@ struct ICCProfile
 	std::vector<uint8_t> getData() const noexcept { return m_Data; };
 
 	/// Return the absolute size of the data
-	uint32_t getDataSize() const noexcept { return m_Data.size(); };
+	uint32_t getDataSize() const noexcept { return static_cast<uint32_t>(m_Data.size()); };
 
 private:
 	std::vector<uint8_t> m_Data;
